fix: updates on socialite token

+++ app/Services/UserTokenService.php
@@ -42,6 +42,10 @@ class UserTokenService
             $user = User::find(Auth::id());
         }
 
+        if (! $user) {
+            return;
+        }
+
         if ($all) {
             $user->tokens()->forceDelete();
